Updated picojson to version 1.3.1-dev.
In order to fix a compilation warning under recent gcc versions.
It also contains a patch which we used to maintain ourselves.

Note that since version 1.3.1 hasn't been released - picojson doesn't
seem actively maintained - this is really just the current master, and
I'm including the short commit hash in the version to indicate exactly
which version we're using.